Refactor scan tasks Add new celery-RMQ Queue: scanmgt
