+++ VERSION
@@ -1,1 +1,1 @@
-1.6.6 // Community Edition
+1.6.7 // Community Edition
