+++ app/settings.py.sample
@@ -286,8 +286,8 @@ BROKER_CONNECTION_TIMEOUT = 30
 # BROKER_POOL_LIMIT = None
 
 # CELERY
-CELERY_RESULT_BACKEND = None
-# CELERY_RESULT_BACKEND = 'django-db'
+#CELERY_RESULT_BACKEND = None
+CELERY_RESULT_BACKEND = 'django-db'
 CELERY_ACCEPT_CONTENT = ['application/json']
 CELERY_TASK_SERIALIZER = 'json'
 CELERY_RESULT_SERIALIZER = 'json'
@@ -315,6 +315,8 @@ SUPERVISORD_API_URL = 'http://localhost:9001/RPC2'
 CELERYFLOWER_API_URL = 'http://localhost:5555/api/'
 
 # More settings
+HTTP_REQUEST_MAX_TIMEOUT = int(os.environ.get('HTTP_REQUEST_MAX_TIMEOUT', 600))
+SCAN_JOB_DEFAULT_MAX_TIMEOUT = int(os.environ.get('SCAN_JOB_DEFAULT_MAX_TIMEOUT', 600))
 ASSETS_MAX = os.environ.get('PATROWL_ASSETS_MAX', 100000000)
 ASSETS_MARGIN = os.environ.get('PATROWL_ASSETS_MARGIN', 10)
 ASSETGROUPS_MAX = os.environ.get('PATROWL_ASSETGROUPS_MAX', 100000000)
