+++ var/etc/supervisord-celery.conf
@@ -47,6 +47,18 @@ stopasgroup=true
 killasgroup=true
 priority=990
 
+[program:celery-scanmgt]
+command=env3/bin/celery worker --hostname=scanmgt-node@%%n --app=app -l info -Q scanmgt --purge --without-mingle --without-gossip --without-heartbeat -Ofair
+stdout_logfile=var/log/celeryd.scanmgt.log
+stderr_logfile=var/log/celeryderr.scanmgt.log
+autostart=true
+autorestart=true
+startsecs=5
+stopwaitsecs=10
+stopasgroup=true
+killasgroup=true
+priority=990
+
 ;[group:celery-workers]
-;programs=celery-flow,celery-beat,celery-default,celery-scan
+;programs=celery-flow,celery-beat,celery-default,celery-scan,celery-scanmgt
 ;priority=990
